pyproject.toml
[project]
name = "ada"
version = "0.1.0"
description = "Another Digital Assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests"
    , "rich"
    , "websockets>=13.1"
    , "pyaudio"
    , "python-dotenv"
    , "ipykernel"
    , "ipywidgets"
    , "openai>=1.51.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
ada = "ada.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
